Reference

Hurst.generalised_hurst_exponent

Hurst.generalised_hurst_range

Hurst.hurst_exponent

Hurst.zeta_estimator_range

Hurst.generalised_hurst_exponent — Method

generalised_hurst_exponent(X, τ_range, q)

Calculate the generalised Hurst exponent of the series X with absolute moment q over the range τ_range along with its standard error.

See also hurst_exponent.

Examples

julia> X = accumulate(+, randn(1000));

julia> generalised_hurst_exponent(X, 1:19, 0.5);

Hurst.generalised_hurst_range — Method

generalised_hurst_range(X, τ_range, q_range)

Calculate the generalised Hurst exponent (GHE) of the series X with absolute moments q_range over the range τ_range, along with its standard error.

Returns a (length(q_range), 2) matrix where the first column contains the values of the GHE and the second column contains the standard errors.

See also hurst_exponent.

Examples

julia> X = accumulate(+, randn(1000));

julia> q_range = 0.:0.1:1.; tau_range = 1:19;

julia> generalised_hurst_range(X, tau_range, q_range);

Hurst.hurst_exponent — Method

hurst_exponent(X, τ_range)

Calculate the Hurst exponent of the series X over the range τ_range along with its standard error.

See Buonocore et al. 2016.

Examples

julia> X = accumulate(+, randn(1000));

julia> isapprox(hurst_exponent(X, 1:19)[1], 0.5, atol = 0.1)
true

Hurst.zeta_estimator_range — Method

zeta_estimator_range(X, τ_range, q_range)

Calculate $\zeta (q)$ that satifies:

$\ln \left(E\left[|X(t+\tau)-X(t)|^q\right] ight)=\zeta(q) \ln (\tau)+\ln (K(q))$

for some series $X(t)$, over the vector q_range.

Returns a (length(q_range), 2) matrix where the first column contains the values of the $\zeta(q)$ for different q and the second column contains the standard errors.

See also hurst_exponent.